pyproject.toml
[project]
name = "luna-bot-discord"
version = "0.1.0"
description = "Um bot de Discord para comunidades de Wild Rift com foco em equipes e torneios."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "dnspython>=2.7.0",
    "motor>=3.7.1",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
]
